requirements.txt
streamlit==1.47.1
pandas==2.3.1
plotly==6.2.0
openpyxl==3.1.5
Pillow==11.3.0